requirements.txt
appnope==0.1.0
attrs==18.2.0
backcall==0.1.0
bleach==3.0.2
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
common==0.1.2
confusable-homoglyphs==3.2.0
coverage==4.5.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-heroku==0.3.1
django-jquery==3.1.0
django-nose==1.4.6
django-registration==3.0
entrypoints==0.2.3
factory-boy==2.11.1
Faker==0.9.2
google-api-core==1.4.1
google-auth==1.5.1
google-cloud-vision==0.34.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
gunicorn==19.9.0
idna==2.7
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.0
numpy==1.15.2
oauthlib==2.1.0
opencv-python==3.4.3.18
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.4.2
prompt-toolkit==2.0.6
protobuf==3.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
pyparsing==2.2.2
pyrsistent==0.14.5
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.20.0
requests-oauthlib==1.0.0
rsa==4.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1
widgetsnbextension==3.4.2
